refactor lumino-1-send.py and lumino-2-send.py into lumino-send.py, which is parametrized
